+++ src/OpenTelemetry/Metrics/MeterProviderSdk.cs
@@ -195,8 +195,8 @@ namespace OpenTelemetry.Metrics
                                 Metric metric;
                                 var metricDescription = metricStreamConfig?.Description ?? instrument.Description;
                                 string[] tagKeysInteresting = metricStreamConfig?.TagKeys;
-                                double[] histogramBucketBounds = (metricStreamConfig is HistogramConfiguration histogramConfig
-                                    && histogramConfig.BucketBounds != null) ? histogramConfig.BucketBounds : null;
+                                double[] histogramBucketBounds = (metricStreamConfig is ExplicitBucketHistogramConfiguration histogramConfig
+                                    && histogramConfig.Boundaries != null) ? histogramConfig.Boundaries : null;
                                 metric = new Metric(instrument, temporality, metricName, metricDescription, histogramBucketBounds, tagKeysInteresting);
 
                                 this.metrics[index] = metric;