+++ src/OpenTelemetry/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Renamed `HistogramConfiguration` to `ExplicitBucketHistogramConfiguration`
+  and changed its member `BucketBounds` to `Boundaries`.
+  ([#2638](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2638))
+
 * Metrics with the same name but from different meters are allowed.
   ([#2634](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2634))